t/slam: Fix compiler warning about string overflow
Explicitly limit the range of the camera numbering when formatting
the name for the debug window to quieten a -Wformat-truncation
warning from the compiler

+++ src/xrt/auxiliary/tracking/t_tracker_slam.cpp
@@ -1078,7 +1078,7 @@ setup_ui(TrackerSlam &t)
 	u_var_add_f32(&t, &t.gravity_correction.z, "Gravity Correction");
 	for (size_t i = 0; i < t.ui_sink.size(); i++) {
 		char label[] = "Camera NNNN";
-		(void)snprintf(label, sizeof(label), "Camera %zu", i);
+		(void)snprintf(label, sizeof(label), "Camera %u", (uint8_t)i);
 		u_var_add_sink_debug(&t, &t.ui_sink[i], label);
 	}
